refactor(examples): rename telemetry example to browser-telemetry
